feat(endpoints): add example task endpoint
The example job endpoint is for example only, demonstrating a workflow
that actively runs jobs given a user's call to a single endpoint (rather
than multiple calls to multiple endpoints) using celery and
mongodb. Effort was taken to minimize the impact on existing code though
integration in the future would be straightforward.

A few key points:

- The API server handles both supplying jobs to the celery worker and
storing the results in the database

- This is meant to be run on each call as opposed to passively as is the
case with Dagster.

- Shouldn't be seen as a replacement for existing (and well documented)
infrastructure, but an alternative to exist alongside it.

- Future work would involve further decoupling of models from the api
endpoints and maybe deeper thought into what endpoint to put future
services like this under (like, `/services` for instance).