feat(models): add task model
This adds a task model based on the ga4gh [Task Execution Service
API](https://ga4gh.github.io/task-execution-schemas/docs/#section/Executive-Summary)
for describing and executing batch execution tasks. While it is not a 1-to-1 mapping of the schema described in the specification it should at least be broadly compatible.