feat(workflow_automation): add activities
We've encapsulated workflow activities within a component that aligns
with the schema and try to build a service that intuitively maps to
it. We've also namespaced this component so that it appears under
`nmdc_runtime/workflow_execution_activity` for easier importing around
the project as well as by external users.