Clean up job adding service for activities
Create jobs for correct workflows

Rewrote the `create_jobs` function to only create new workflows if they
do not have an activity already associated with them in the incoming
workflow execution activity sets. This method creates a mapping of nodes
and their children and then iterates through them to see if their
children account for all of the workflow execution activities that
should have been created already. If something is missing a new job will
be created for it.