fix: refresh minter.requesters collection on site creation
The interface to modifying the sites collection is limited to a single endpoint,
so I think it is fine to directly call a function to perform the operation.

I can envision moving to a more event-driven handler architecture once enough of these cases are present.
I don't think it's worth the added complexity at this time.

fixes #317