Remove buggy expand/collapse functionality within metadata descriptions
The implementation contained a bug: it relied on a `<details />` element
that accepted user clicks; but was within an element that was already
listening for user clicks (i.e. the metadata description element).
Clicking on the `<details />` element also counted as clicking on all
of its parent elements (there was no `event.stopPropagation()` call).
As a result, when the `<details />` element was expanded, the endpoints
would be collapsed—and vice versa.