Bump `nmdc-schema` from `11.3.0` to `11.4.0` without updating Dagster
In this commit, I bumped `nmdc-schema` in `main.in`, then ran:
```
make update-deps \
UPDATE_DEPS_MAIN_UPGRADE_OPT='' UPDATE_DEPS_DEV_UPGRADE_OPT=''
```
When those two options are included, the underlying `pip-compile`
command does not include the `--upgrade` option. As a result,
`pip-compile` does not try to upgrade all dependencies to their
latest versions.

I did this because, when the `--upgrade` option was being passed to
`pip-compile`, the automated test suite in GitHub Actions was failing
due to something related to Dagster, which did get updated in a way
that has breaking changes documented in its changelog. This is my
attempt to avoid that for now (kick the can down the road).