Revert nested tag functionality and have search use deep linking
Nested tag (hierarchical tag) functionality was not working with
deep linking, and "scrollTo" functionality was not working for
endpoints that were within collapsed sections.

With nested tag functionality out of the picture, Swagger UI
automatically handles deep linking, which automatically
handles expanding collapsed sections as necessary.

I think this is an overall better solution. Still to do:
reorganize the endpoints and refine the group names.